1) Increase minimun supported android API to 8. As shown in http://developer.android.com/about/dashboards/index.html beginning in September, 2013, Play Store don't support API <= 7. Devices with API less than 8 are around 1% of total. Devices with API level 8 are around 1% of total in January 2014.
    2) Cleanup and solve easy code warnings.

+++ OsmAnd-java/src/net/osmand/binary/BinaryMapRouteReaderAdapter.java
@@ -525,7 +525,7 @@ public class BinaryMapRouteReaderAdapter {
 					fromr.restrictions = new long[it.value().size()];
 					for (int k = 0; k < fromr.restrictions.length; k++) {
 						int to = (int) (it.value().get(k) >> RouteDataObject.RESTRICTION_SHIFT);
-						long valto = (idTables.get(to) << RouteDataObject.RESTRICTION_SHIFT) | ((long) it.value().get(k) & RouteDataObject.RESTRICTION_MASK);
+						long valto = (idTables.get(to) << RouteDataObject.RESTRICTION_SHIFT) | (it.value().get(k) & RouteDataObject.RESTRICTION_MASK);
 						fromr.restrictions[k] = valto;
 					}
 				}
